+++ src/activity/Activity.js
@@ -143,15 +143,7 @@ Object.defineProperties(Activity.prototype, {
       let formatter = this[kFormatter];
       if (formatter) return formatter;
 
-      const broker = this.broker;
-      formatter = this[kFormatter] = new Formatter(
-        {
-          id: this.id,
-          broker,
-          logger: this.logger,
-        },
-        broker.getQueue('format-run-q')
-      );
+      formatter = this[kFormatter] = new Formatter(this);
       return formatter;
     },
   },
@@ -430,7 +422,12 @@ Activity.prototype._discardRun = function discardRun() {
   const execution = this[kExec].get('execution');
   if (execution && !execution.completed) return;
 
+  let discardRoutingKey = 'run.discard';
   switch (status) {
+    case 'executed': {
+      discardRoutingKey = 'run.discarded';
+      break;
+    }
     case 'end':
     case 'executing':
     case 'error':
@@ -442,10 +439,11 @@ Activity.prototype._discardRun = function discardRun() {
 
   const stateMessage = this[kStateMessage];
   if (this.extensions) this.extensions.deactivate(cloneMessage(stateMessage));
+
   const broker = this.broker;
   broker.getQueue('run-q').purge();
 
-  broker.publish('run', 'run.discard', cloneContent(stateMessage.content));
+  broker.publish('run', discardRoutingKey, cloneContent(stateMessage.content), { correlationId: stateMessage.properties.correlationId });
   this[kConsuming] = true;
   this._consumeRunQ();
 };
@@ -602,12 +600,13 @@ Activity.prototype._onRunMessage = function onRunMessage(routingKey, message, me
 
   const preStatus = this.status;
   this.status = 'formatting';
-  console.log({ routingKey, mc: this.formatter.formatQ.messageCount });
+
   return this.formatter.format(message, (err, formattedContent, formatted) => {
-    console.log({ routingKey, formatted, mc: this.formatter.formatQ.messageCount });
-    if (err) return this.emitFatal(err, message.content);
-    if (formatted) message.content = formattedContent;
     this.status = preStatus;
+    if (err) {
+      return this.emitFatal(err, message.content);
+    }
+    if (formatted) message.content = formattedContent;
     this._continueRunMessage(routingKey, message, messageProperties);
   });
 };